requirements-test.txt
ansible-core==2.16.2
ansible-lint
molecule
molecule-docker
influxdb_client